requirements.txt
numpy<2.0.0
torch
opencv-python
pillow
ollama
